% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METAL_gwas.R
\name{METAL_gwas}
\alias{METAL_gwas}
\title{使用METAL软件对GWAS进行meta分析}
\usage{
METAL_gwas(
  GWASfile = c("./dat1.txt", "./dat1.txt"),
  Metal_path = "./metal.exe",
  GWAS_name = "METAL_gwas",
  save_path = "./",
  type = "outcome",
  Analysis_Scheme = "samplesize",
  heterogeneity = TRUE,
  Average_FREQ = TRUE,
  MinMax_FREQ = FALSE,
  Filter_N = FALSE,
  Filter_MAF = FALSE,
  Genomic_Control_Correction = FALSE,
  Sample_Overlap_Correction = FALSE
)
}
\arguments{
\item{GWASfile}{字符向量，包含需要meta分析的所有TwoSampleMR格式的完整GWAS文件路径，注意必须是相同type类型。}

\item{Metal_path}{METAL软件的可执行文件路径，例如"./metal.exe"。}

\item{GWAS_name}{输出文件的文件名。}

\item{save_path}{文件保存路径。}

\item{type}{TwoSampleMR格式数据类型，outcome或exposure。}

\item{Analysis_Scheme}{指定分析加权方法，samplesize或者SE。samplesize表示基于权重（通常为样本量）和效应方向，跨研究合并p值。
SE表示使用相应标准误差的倒数对效应大小估计值进行加权。}

\item{heterogeneity}{TRUE或FALSE，是否进行异质性分析。}

\item{Average_FREQ}{TRUE或FALSE，是否计算平均频率。}

\item{MinMax_FREQ}{TRUE或FALSE，是否计算最小和最大频率。}

\item{Filter_N}{数值或FALSE，如果为数值，则添加一个基于样本数量的过滤。如果为FALSE，则不添加此过滤。}

\item{Filter_MAF}{数值或FALSE，如果为数值，则添加一个基于次要等位基因频率(MAF)的过滤值，一般设置为0.01。
如果为FALSE，则不添加此过滤。}

\item{Genomic_Control_Correction}{TRUE或FALSE，是否进行基因组控制校正。}

\item{Sample_Overlap_Correction}{TRUE或FALSE，是否进行样本重叠校正。}
}
\description{
windows系统下使用METAL软件对全基因组关联研究(GWAS)数据进行META分析，
注意：输入的数据格式必须为完整的TwoSampleMR格式文件，电脑内存小容易运行崩溃，
且必须包含SNP、samplesize、effect_allele、other_allele、eaf、beta、se、pval列，
且数据不能有缺失，若需要MAF过滤，需提前将eaf转换，并包含MAF列。
}
